Updated project dependencies, configuration, and Kubernetes setup
This commit includes a wide range of updates across the project. Key changes include updating various package versions in multiple .csproj files and modifying project configurations in appsettings.json and appsettings.Development.json. The Dockerfile was updated to use new images for building and running the application. Several Kubernetes configuration files were added and updated in kubernetes.dcproj. The .gitignore file was updated to ignore several new files. The project type for WebApps.Backoffice.UnitTests was changed in OmniRetailSuite.sln. Additionally, some code refactoring was done, including the replacement of the IdentityUriHelper class with the IdentityUriResolver class in UserPreferenceService.cs.