Refactor methods and update property assignments in CashiersWorkplace and Receipt classes
In this commit, several methods in `CashiersWorkplace.razor.cs` have been refactored. The `OnInitializedAsync` method has been replaced with `OnInitialized`, and `OnInitializedReceipt` has been replaced with `InitializeReceiptCommands`, which does not return a Task. The `InitNewReceipt` method call has been moved from `InitializeReceiptCommands` to `OnAfterRenderAsync`. The `SaveReceipt` method no longer calls `ReceiptService.CreateAsync(_receipt)`, instead it now calls `CallRequestRefresh()`. The `CashierOnSelectionChanged` method has been updated to return if the `selectedCashier` is null, and if not, it sets `_cashier` and `_receipt.Cashier` to `selectedCashier`.

In `Receipt.cs`, `Store` and `Cashier` properties now set their respective Ids when the value is not null, otherwise the Id is set to `Guid.Empty`. The `TryUpdateReceiptItemByCatalogItem` method has been updated to return false if `receiptItem` is null, and if not, it updates `receiptItem.Quantity` and `receiptItem.TotalPrice`. The `CreateReceiptItemByCatalogProductItem` method no longer sets `ProductItemId`.

In `ReceiptItem.cs`, `ProductItem` property now sets `ProductItemId` when the value is not null, otherwise `ProductItemId` is set to `Guid.Empty`.