Added ContextMenu component and refactored ItemList
A new `ContextMenu` component was added, encapsulating the functionality of a context menu. This component uses the `MudMenu` component and a new `ContextMenuItem` class to define menu items. The `ContextMenu` component was integrated into the `ItemList` component, replacing the previous usage of `MudMenu`.

Several changes were made in `ItemList.razor.cs` to accommodate this new component, including the removal of `IJSRuntime` injection and related code, the addition of a `DefineContextMenuItems` method, and refactoring of the `OpenItemClick` method. The `SetFilterString` method was also removed