Add "doc" project and update architecture diagram
Added a new project named "doc" to the OmniRetailSuite solution, which includes a solution item named "High-Level Architecture Overview.drawio". Updated the "High-Level Architecture Overview.drawio" file, a diagram that outlines the high-level architecture of the system. The diagram now includes various components and databases, as well as frontend and backend services.