Updated `GetReceiptAsync` for eager loading of related data
Updated the `GetReceiptAsync` method in the `ReceiptRepository` class to include related data in the query results. This includes the `Cashier`, `Store`, and `ReceiptItems` related to each `Receipt`, as well as the `ProductItem` related to each `ReceiptItem`. This is achieved using the `Include` and `ThenInclude` methods from Entity Framework Core, reducing the need for additional database queries.