Updated namespaces, mappings, and form layout in Retail app
In this commit, several changes were made to the Retail application. The `Retail.Application.Configuration` namespace was added to `Program.cs` and replaced the old namespace in `ConfigureAppServices.cs`. New mappings were added in `ReceiptItemProfile.cs` and `ReceiptProfile.cs` for `ReceiptItemDto` to `ReceiptItem` and `ReceiptDto` to `Receipt` respectively. Lastly, the layout of the form in `ReceiptDetails.razor` was updated, moving "Store" and "Cashier" fields to a new `MudCard` and adding a "Total" field to the first `MudCard`.