Revamp loading screen with new design and styles
Removed old SVG-based loading progress indicator and associated text from index.html. Added a new loading screen within a div with the class `ors-loader`, which includes an image, a heading, and a spinner with a loading message. Updated app.css to remove `.loading-progress` and `.loading-progress-text` classes and added new styles for `.ors-loader` and `#ors-loader img` to enhance the user experience during the application's loading phase.